Move db schema closer to adapter
