+++ app/db/registry.py
@@ -1,5 +1,5 @@
 from app.bill.schemas import Bill  # noqa:F401
 from app.order.schemas import Order, OrderItem  # noqa:F401
 from app.payment.schemas import Payment  # noqa:F401
-from app.product.schemas import Product, ProductVariant  # noqa:F401
+from app.product.adapters.sql import Product, ProductVariant  # noqa:F401
 from app.user.schemas import User, UserCredentials  # noqa:F401
